Refactor container lookup stuff into an instanceIntializer
